refactor(gateway-contracts): rename task file for FHE keys gen
