+++ gateway-contracts/hardhat.config.ts
@@ -16,9 +16,9 @@ import "./tasks/blockExplorerVerify";
 import "./tasks/deployment/contracts";
 import "./tasks/deployment/empty_proxies";
 import "./tasks/deployment/mock_contracts";
+import "./tasks/generateFheKeys";
 import "./tasks/getters";
 import "./tasks/safeSmartAccounts";
-import "./tasks/triggerKeygen";
 import "./tasks/upgradeContracts";
 
 const dotenvConfigPath: string = process.env.DOTENV_CONFIG_PATH || "./.env";
